+++ services/lomax/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for expansion of `{hostname}` to the real hostname in the
+  `certificate_alt_names` config setting.
+
 ## [2.0.0] - 2021-12-16
 
 ### Added